End2end: Improved assertion msgs and flattened assertion funcs
Details:

- Created a new module assertions.py and moved all assertion
  functions from the ProfileTest class to the new module.
  This allows using them in more ways than just for testcases
  that use ProfileTest.

- Improved all assertion messages of end2end tests so that the
  server and the test context is now always part of the message.
  In order to do that, all assertion functions that did not have
  that yet, got the WBEMConnection object added as a new parameter.
  Also as part of that, the WBEMConnection objects used for the
  end2end tests got a new attribute 'server_definition'
  dynamically added, that allows getting the server definition
  nickname for the assertion messages, which is more recognizable
  than the server URL. The URL is still also part of the message.

- Changed some assert statements into raising AssertionError,
  in order to have more control about the assertion message.
